+++ heylogs-api/src/main/java/nbbrd/heylogs/Config.java
@@ -9,9 +9,8 @@ public class Config {
 
     public static final Config DEFAULT = Config.builder().build();
 
-    @NonNull
-    @lombok.Builder.Default
-    String versionTagPrefix = "";
+    @Nullable
+    String versionTagPrefix;
 
     @Nullable
     String versioningId;