+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add Calendar Versioning [#253](https://github.com/nbbrd/heylogs/issues/253)
+- Add forge configuration [#421](https://github.com/nbbrd/heylogs/issues/421)
 
 ### Changed